move handlers to a different place in the codebase
