+++ api/routes/push.js
@@ -1,64 +1,11 @@
 const router = require('express').Router();
-const notifications = require('../notifications/notifications');
-const middleware = require('./middleware');
+const pushHandlers = require('../handlers/push');
+const middleware = require('../handlers/middleware');
 
-// import UserSettings schema
-let UserSettings = require('../models/userSettings');
+// set an fcm token for authenticated user
+router.post('/fcmToken', middleware.verifyToken, pushHandlers.post);
 
-router.post('/fcmToken', middleware.verifyToken, async function(req, res) {
-    try {
-        const userId = req.authorization.id;
-        const userFcmToken = req.body.fcmToken;
-
-        if (!userFcmToken) {
-          return res.status(400).send("No fcm token supplied");
-        }
-
-        let userSettings = await UserSettings.findById(userId);
-
-        userSettings.fcmToken = userFcmToken;
-        await userSettings.save();
-
-        // return the received fcm token as confirmation
-        res.json({token: userFcmToken});
-    }
-    catch (e) {
-        console.log(`[error] ${e}`);
-        res.status(500).send("Error storing FCM Token");
-    }
-});
-
-// test push notification endpoint
-router.post('/push', middleware.verifyToken, async function(req, res) {
-    try {
-        const uid = req.body.userId;
-        if (!uid) {
-          return res.status(400).send("no user id specified");
-        }
-
-        let userToNotify = await UserSettings.findById(uid);
-
-        if (!userToNotify) {
-          return res.status(400).send("user does not exist");
-        }
-
-        if (!userToNotify.fcmToken) { // should never happen
-          console.log(`tried to send notification to ${userToNotify.name} but they had no valid FCM token`);
-          return res.status(400).send("user does not have valid fcm token");
-        }
-
-        let notification = {
-            title: "New Test Notification!",
-            body: `You received a test push notification from ${req.authorization.email}`
-        }
-        notifications.notify(notification, [userToNotify.fcmToken]);
-
-        res.json({success: true, message: "user notified"});
-    }
-    catch (e) {
-        console.log(`[error] ${e}`);
-        res.status(500).send("Error sending push notification");
-    }
-});
+// test push notification token for a user
+router.post('/push', middleware.verifyToken, pushHandlers.test);
 
 module.exports = router;
